# pkg/api/v1/docs/api_swagger.yaml
basePath: /v1
consumes:
- application/json
definitions:
...
  models.ClientTemplateResponse:
    properties:
      allow_additional_dns_names:
        type: boolean
      allow_additional_ips:
        type: boolean
      authority_name:
        type: string
      common_name:
        type: string
      created_at:
        type: string
      dns_names:
        items:
          type: string
        type: array
      id:
        type: string
      ip_addresses:
        items:
          type: string
        type: array
      name:
        type: string
      tag:
        type: string
      validity:
        type: string
    type: object
...
  models.CreateClientTemplateRequest:
    properties:
      allow_additional_dns_names:
        type: boolean
      allow_additional_ips:
        type: boolean
      authority_name:
        type: string
      common_name:
        type: string
      dns_names:
        items:
          type: string
        type: array
      ip_addresses:
        items:
          type: string
        type: array
      name:
        type: string
      tag:
        type: string
      validity:
        type: string
    type: object
...
  models.DeleteClientTemplateRequest:
    properties:
      authority_name:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Returns the health and status of the various services that make up
    the API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: EndKey API V1
  version: "1.0"
paths:
...
  /template/client:
    delete:
      consumes:
      - application/json
      description: Delete a Client Template
      parameters:
      - description: Delete Client Template Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteClientTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - template
    post:
      consumes:
      - application/json
      description: Create a new Client Template
      parameters:
      - description: Create Client Template Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateClientTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ClientTemplateResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - template
  /template/client/{authority_name}:
    get:
      consumes:
      - application/json
      description: List all Client Templates
      parameters:
      - description: Authority Name
        in: path
        name: authority_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ClientTemplateResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
        "412":
          description: Precondition Failed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      tags:
      - template
...
produces:
- application/json
schemes:
- https
swagger: "2.0"